+++ src/types.d.ts
@@ -1,3 +1,1 @@
 declare module '*.png';
-
-type Theme = 'dark' | 'light';
